fix: fix previous url context to store all previous urls
